separated platform files
